browser/devtools/webide/test/test_device_permissions.html

<!DOCTYPE html>

<html>

  <head>
    <meta charset="utf8">
    <title></title>

    <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
    <script type="application/javascript" src="chrome://mochikit/content/chrome-harness.js"></script>
    <script type="application/javascript;version=1.8" src="head.js"></script>
    <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  </head>

  <body>

    <script type="application/javascript;version=1.8">
      window.onload = function() {
        SimpleTest.waitForExplicitFinish();

        Task.spawn(function* () {
          const { DebuggerServer } = require("devtools/server/main");
          Services.prefs.setBoolPref("devtools.webide.sidebars", false);

          if (!DebuggerServer.initialized) {
            DebuggerServer.init();
            DebuggerServer.addBrowserActors();
          }

          let win = yield openWebIDE();

          let permIframe = win.document.querySelector("#deck-panel-permissionstable");

          yield connectToLocalRuntime(win);

          let perm = win.document.querySelector("#cmd_showPermissionsTable");

          ok(!perm.hasAttribute("disabled"), "perm cmd enabled");

          let deck = win.document.querySelector("#deck");

          win.Cmds.showPermissionsTable();
          is(deck.selectedPanel, permIframe, "permission iframe selected");

          yield nextTick();

          yield lazyIframeIsLoaded(permIframe);

          yield permIframe.contentWindow.getRawPermissionsTablePromise;

          doc = permIframe.contentWindow.document;
          trs = doc.querySelectorAll(".line");
          found = false;
          for (let tr of trs) {
            let [name,v1,v2,v3] = tr.querySelectorAll("td");
            if (name.textContent == "geolocation") {
              found = true;
              is(v1.className, "permprompt", "geolocation perm is valid");
              is(v2.className, "permprompt", "geolocation perm is valid");
              is(v3.className, "permprompt", "geolocation perm is valid");
              break;
            }
          }
          ok(found, "Found geolocation line");

          doc.querySelector("#close").click();

          ok(!deck.selectedPanel, "No panel selected");

          DebuggerServer.destroy();

          yield closeWebIDE(win);

          SimpleTest.finish();
        }).then(null, e => {
          ok(false, "Exception: " + e);
          SimpleTest.finish();
        });
      }
    </script>
  </body>
</html>
